refactor: Restructuring autoeval module
Core functionality is now kept in the core module
